overlay toolbars (#435)
* init impl

* improve styling

* overlays css tweaks
