+++ rnote-ui/data/ui/appwindow.ui
@@ -39,107 +39,6 @@
               </object>
             </child>
             <child>
-              <object class="GtkBox">
-                <property name="orientation">vertical</property>
-                <property name="hexpand">true</property>
-                <property name="halign">fill</property>
-                <layout>
-                  <property name="column">0</property>
-                  <property name="row">2</property>
-                  <property name="column-span">3</property>
-                </layout>
-                <style>
-                  <class name="topbar_box" />
-                </style>
-                <child>
-                  <object class="GtkFlowBox">
-                    <property name="orientation">horizontal</property>
-                    <property name="selection-mode">none</property>
-                    <property name="hexpand">true</property>
-                    <property name="vexpand">false</property>
-                    <property name="halign">fill</property>
-                    <property name="valign">start</property>
-                    <property name="margin-start">6</property>
-                    <property name="margin-end">6</property>
-                    <child>
-                      <object class="GtkBox" id="pens_toggles_box">
-                        <property name="hexpand">true</property>
-                        <property name="halign">center</property>
-                        <property name="valign">center</property>
-                        <property name="width-request">400</property>
-                        <property name="homogeneous">true</property>
-                        <style>
-                          <class name="toolbar" />
-                        </style>
-                        <child>
-                          <object class="GtkToggleButton" id="brush_toggle">
-                            <property name="icon_name">pen-brush-symbolic</property>
-                            <property name="tooltip_text" translatable="yes">Brush</property>
-                            <property name="hexpand">true</property>
-                          </object>
-                        </child>
-                        <child>
-                          <object class="GtkToggleButton" id="shaper_toggle">
-                            <property name="icon_name">pen-shaper-symbolic</property>
-                            <property name="tooltip_text" translatable="yes">Shaper</property>
-                            <property name="hexpand">true</property>
-                            <property name="group">brush_toggle</property>
-                          </object>
-                        </child>
-                        <child>
-                          <object class="GtkToggleButton" id="typewriter_toggle">
-                            <property name="icon_name">pen-typewriter-symbolic</property>
-                            <property name="tooltip_text" translatable="yes">Typewriter</property>
-                            <property name="hexpand">true</property>
-                            <property name="group">brush_toggle</property>
-                          </object>
-                        </child>
-                        <child>
-                          <object class="GtkToggleButton" id="eraser_toggle">
-                            <property name="icon_name">pen-eraser-symbolic</property>
-                            <property name="tooltip_text" translatable="yes">Eraser</property>
-                            <property name="hexpand">true</property>
-                            <property name="group">brush_toggle</property>
-                          </object>
-                        </child>
-                        <child>
-                          <object class="GtkToggleButton" id="selector_toggle">
-                            <property name="icon_name">pen-selector-symbolic</property>
-                            <property name="tooltip_text" translatable="yes">Selector</property>
-                            <property name="hexpand">true</property>
-                            <property name="group">brush_toggle</property>
-                          </object>
-                        </child>
-                        <child>
-                          <object class="GtkToggleButton" id="tools_toggle">
-                            <property name="icon_name">pen-tools-symbolic</property>
-                            <property name="tooltip_text" translatable="yes">Tools</property>
-                            <property name="hexpand">true</property>
-                            <property name="group">brush_toggle</property>
-                          </object>
-                        </child>
-                      </object>
-                    </child>
-                    <child>
-                      <object class="ColorPicker" id="colorpicker">
-                        <property name="amount-colorbuttons">8</property>
-                        <property name="hexpand">true</property>
-                        <property name="halign">center</property>
-                        <property name="valign">center</property>
-                        <property name="position">top</property>
-                        <property name="margin-top">3</property>
-                        <property name="margin-bottom">3</property>
-                      </object>
-                    </child>
-                  </object>
-                </child>
-                <child>
-                  <object class="GtkSeparator">
-                  </object>
-                </child>
-              </object>
-            </child>
-            <child>
               <object class="RnoteOverlays" id="overlays">
                 <property name="hexpand">true</property>
                 <property name="vexpand">true</property>
