+++ rnote-ui/data/ui/mainheader.ui
@@ -28,13 +28,98 @@
         <property name="centering-policy">loose</property>
         <property name="vexpand">false</property>
         <child type="start">
-          <object class="GtkBox">
-            <property name="spacing">3</property>
+          <object class="GtkBox" id="quickactions_box">
+            <property name="spacing">6</property>
             <child>
               <object class="GtkButton" id="create_page_button">
                 <property name="icon-name">tab-new-symbolic</property>
                 <property name="action-name">win.new-tab</property>
                 <property name="tooltip-text" translatable="yes">New tab</property>
+              </object>
+            </child>
+            <child>
+              <object class="GtkRevealer" id="fixedsize_quickactions_revealer">
+                <property name="transition-type">slide-right</property>
+                <property name="transition-duration">300</property>
+                <property name="overflow">visible</property>
+                <child>
+                  <object class="GtkBox">
+                    <property name="homogeneous">true</property>
+                    <property name="spacing">6</property>
+                    <child>
+                      <object class="GtkButton">
+                        <property name="icon_name">add-page-symbolic</property>
+                        <property name="tooltip_text" translatable="yes">Add page</property>
+                        <property name="action-name">win.add-page-to-doc</property>
+                      </object>
+                    </child>
+                    <child>
+                      <object class="GtkButton">
+                        <property name="icon_name">resize-to-fit-strokes-symbolic</property>
+                        <property name="tooltip_text" translatable="yes">Resize document to fit
+                          the
+                          strokes</property>
+                        <property name="action-name">win.resize-to-fit-strokes</property>
+                      </object>
+                    </child>
+                  </object>
+                </child>
+              </object>
+            </child>
+            <child>
+              <object class="GtkBox">
+                <style>
+                  <class name="linked" />
+                </style>
+                <child>
+                  <object class="GtkButton">
+                    <property name="icon_name">zoom-out-symbolic</property>
+                    <property name="tooltip_text" translatable="yes">Zoom out</property>
+                    <property name="action-name">win.zoom-out</property>
+                    <style>
+                      <class name="flat" />
+                    </style>
+                  </object>
+                </child>
+                <child>
+                  <object class="GtkButton">
+                    <property name="icon_name">zoom-in-symbolic</property>
+                    <property name="tooltip_text" translatable="yes">Zoom in</property>
+                    <property name="action-name">win.zoom-in</property>
+                    <style>
+                      <class name="flat" />
+                    </style>
+                  </object>
+                </child>
+              </object>
+            </child>
+            <child>
+              <object class="GtkBox">
+                <style>
+                  <class name="linked" />
+                </style>
+                <child>
+                  <object class="GtkButton" id="undo_button">
+                    <property name="sensitive">false</property>
+                    <property name="icon_name">edit-undo-symbolic</property>
+                    <property name="tooltip_text" translatable="yes">Undo</property>
+                    <property name="action-name">win.undo</property>
+                    <style>
+                      <class name="flat" />
+                    </style>
+                  </object>
+                </child>
+                <child>
+                  <object class="GtkButton" id="redo_button">
+                    <property name="sensitive">false</property>
+                    <property name="icon_name">edit-redo-symbolic</property>
+                    <property name="tooltip_text" translatable="yes">Redo</property>
+                    <property name="action-name">win.redo</property>
+                    <style>
+                      <class name="flat" />
+                    </style>
+                  </object>
+                </child>
               </object>
             </child>
           </object>
